handle comments
