+++ packages/api-providers/src/ipfs/substrate/vanchor.ts
@@ -1,0 +1,12 @@
+import { getCachedFixtureURI, withLocalFixtures } from '@webb-dapp/api-providers/utils';
+
+export async function fetchSubstrateVAnchorProvingKey() {
+  const IPFSUrl = 'https://ipfs.io/ipfs/QmZiNuAKp2QGp281bqasNqvqccPCGp4yoxWbK8feecefML';
+  const cachedURI = getCachedFixtureURI('proving_key_uncompressed_sub_vanchor_2_2_2.bin');
+  const ipfsKeyRequest = await fetch(withLocalFixtures() ? cachedURI : IPFSUrl);
+  const circuitKeyArrayBuffer = await ipfsKeyRequest.arrayBuffer();
+
+  const circuitKey = new Uint8Array(circuitKeyArrayBuffer);
+
+  return circuitKey;
+}
